# docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: TaskFlow API
  version: 1.0.0
  description: |
    TaskFlow 后端 REST API。
    
    ## 鉴权说明
    
    - 普通用户接口：在请求头中携带 `Authorization: Bearer <JWT>`
    - 管理员控制平面接口（`/admin/*`）：在请求头中携带 `Authorization: Bearer <ADMIN_TOKEN>`
    - 两套鉴权体系完全独立
    
    ## 错误格式
    
    所有错误响应统一格式：
    ```json
    { "error": "错误描述", "code": "ERROR_CODE" }
    ```
    
    ## 时间格式
    
    所有时间字段均为 ISO 8601 格式，UTC 时区，例如 `2025-06-01T12:00:00.000Z`。
servers:
  - url: http://localhost:3001
    description: 本地开发环境
  - url: https://api.taskflow.example.com
    description: 生产环境

# ─── Security Schemes ────────────────────────────────────────────────────────
components:
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用户 JWT Token，通过 /auth/login 获取
    adminAuth:
      type: http
      scheme: bearer
      description: 管理员 ADMIN_TOKEN，通过环境变量配置

  # ─── Reusable Schemas ──────────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: 人类可读的错误描述
        code:
          type: string
          description: 机器可读的错误代码
      example:
        error: "资源不存在"
        code: "NOT_FOUND"

    # ── Auth ──
...
    TaskSummary:
      type: object
      required: [id, classId, title, createdAt]
      properties:
        id:
          type: string
          format: uuid
        classId:
          type: string
          format: uuid
        className:
          type: string
        title:
          type: string
        startAt:
          type: string
          format: date-time
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        allowLateSubmission:
          type: boolean
        blockedBy:
          type: array
          items:
            type: string
            format: uuid
        createdBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userState:
          $ref: '#/components/schemas/TaskUserState'
          description: 当前用户对此任务的状态，可能为 null（未产生任何互动时）
          nullable: true
        submittedCount:
          type: integer
          description: 该任务的已提交人数（所有成员可见）
        memberCount:
          type: integer
          description: 该任务所属班级的成员总数（所有成员可见）
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/TaskSummary'
        - type: object
          properties:
            description:
              type: string
              nullable: true
              description: Markdown 正文
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/AttachmentMeta'
            stats:
              $ref: '#/components/schemas/TaskStats'
              description: 仅 OWNER 和 ADMIN 可见
              nullable: true
    TaskStats:
      type: object
      description: 任务的已读/已提交统计，仅管理员可见
      properties:
        memberCount:
          type: integer
        viewedCount:
          type: integer
        submittedCount:
          type: integer
    TaskUserState:
      type: object
      properties:
        viewedAt:
          type: string
          format: date-time
          nullable: true
        tags:
          type: array
          items:
            type: string
        sortOrder:
          type: number
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        startAt:
          type: string
          format: date-time
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        allowLateSubmission:
          type: boolean
        blockedBy:
          type: array
          items:
            type: string
            format: uuid
    UpdateTaskStateRequest:
      type: object
      description: 用户更新自己对某任务的个人状态（tags、排序）
      properties:
        tags:
          type: array
          items:
            type: string
        sortOrder:
          type: number
    ParseTaskRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: 用户输入的自然语言任务描述
    ParseTimeOption:
      type: object
      properties:
        startAt:
          type: string
          format: date-time
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
    ParseTaskResponse:
      type: object
      description: LLM 解析结果。timeOptions 包含 1–3 个可能的时间解释，当 AI 对用户输入的时间表达不确定时返回多个选项，前端展示让用户选择。
      properties:
        title:
          type: string
          nullable: true
        timeOptions:
          type: array
          items:
            $ref: '#/components/schemas/ParseTimeOption'
          minItems: 1
          maxItems: 3
          description: 1–3 个可能的起止时间选项。优先返回 1 个，仅在用户输入有歧义时返回多个。
        allowLateSubmission:
          type: boolean
          nullable: true
          description: 仅当用户明确提及时返回 true/false，否则 null
        description:
          type: string
          nullable: true
    ParseDraftTaskResponse:
      allOf:
        - $ref: '#/components/schemas/ParseTaskResponse'
        - type: object
          properties:
            markdownCached:
              type: boolean
              description: 是否已生成并缓存 markdown 正文草稿

    # ── Submissions ──
    SubmissionListRow:
      type: object
      description: 班级成员提交情况行（管理员视图）
      required: [userId, email, role, submitted, attachments]
      properties:
        userId:
          type: string
          format: uuid
        nickname:
          type: string
          nullable: true
        email:
          type: string
        schoolName:
          type: string
          nullable: true
        studentId:
          type: string
          nullable: true
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER]
        submitted:
          type: boolean
        submission:
          $ref: '#/components/schemas/SubmissionSummary'
          nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentMeta'
          description: 该提交的附件列表（未提交者为空数组）
    SubmissionSummary:
      type: object
      required: [id, taskId, userId, firstSubmittedAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userNickname:
          type: string
          nullable: true
        userEmail:
          type: string
        firstSubmittedAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        score:
          type: number
          nullable: true
        reviewerId:
          type: string
          format: uuid
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reviewNote:
          type: string
          nullable: true
    SubmissionDetail:
      allOf:
        - $ref: '#/components/schemas/SubmissionSummary'
        - type: object
          properties:
            content:
              type: string
              nullable: true
              description: Markdown 正文
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/AttachmentMeta'
    UpsertSubmissionRequest:
      type: object
      properties:
        content:
          type: string
          nullable: true
          description: Markdown 正文
    GradeSubmissionRequest:
      type: object
      properties:
        score:
          type: number
          nullable: true
        reviewNote:
          type: string
          nullable: true

    # ── Attachments ──
    AttachmentMeta:
      type: object
      required: [id, fileKey, originalName, createdAt]
      properties:
        id:
          type: string
          format: uuid
        fileKey:
          type: string
        originalName:
          type: string
        renamedFile:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          nullable: true
        url:
          type: string
          description: 文件访问地址，指向 GET /files/:fileKey
        createdAt:
          type: string
          format: date-time

    # ── Admin ──
...
paths:
...
  /tasks/parse:
    post:
      tags: [Tasks]
      summary: 用 AI 解析自然语言为任务字段
      description: |
        将用户输入的自然语言文本发送给 LLM，返回结构化字段建议。
        
        此接口**不创建任务**，仅做解析。
        
        LLM 返回不确定的字段时，对应值为 null。
        若 LLM 调用失败或返回格式异常，接口仍返回 200，所有字段均为 null（降级处理，不报错）。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseTaskRequest'
      responses:
        '200':
          description: 解析结果（部分或全部字段可能为 null）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseTaskResponse'
  /tasks/{taskId}/parse:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Tasks]
      summary: AI 解析草稿任务（含附件）
      description: |
        对已有草稿任务调用 LLM 解析。会读取任务已上传的附件一并发送给 AI。
        解析结果自动更新草稿的 title、startAt、dueAt、description 字段（使用第一个 timeOption）。
        同时生成 markdown 正文缓存（通过 GET /tasks/{taskId}/draft-markdown 读取）。
        仅 OWNER/ADMIN 可调用。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: 用户输入的自然语言任务描述（可选，不传则使用任务已有的 sourceText）
      responses:
        '200':
          description: 解析结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseDraftTaskResponse'
  /tasks/{taskId}/draft-markdown:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Tasks]
      summary: 读取草稿任务的 AI 生成 markdown 缓存
      description: |
        返回 AI 解析时生成的 markdown 正文草稿（缓存在 Redis，TTL 3 天）。
        仅 OWNER/ADMIN 可调用。
      security:
        - userAuth: []
      responses:
        '200':
          description: markdown 缓存内容
          content:
            application/json:
              schema:
                type: object
                properties:
                  markdown:
                    type: string
                    nullable: true
  /tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Tasks]
      summary: 获取任务详情
      description: |
        返回任务完整信息，包含正文和附件。
        
        管理员额外可见 stats（已读/已提交统计）。
        
        **此接口不自动记录已读**，已读由单独的 POST /tasks/:taskId/view 触发。
      security:
        - userAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 任务不存在或已删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Tasks]
      summary: 更新任务信息
      description: 仅 OWNER 和 ADMIN 可操作。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Tasks]
      summary: 删除任务
      description: |
        仅 OWNER 和 ADMIN 可操作。
        
        - 若任务有零条提交：硬删除，完整清理所有关联数据和 MinIO 文件。
        - 若任务有提交记录：软删除，清空正文和附件，保留任务行供提交引用。
      security:
        - userAuth: []
      responses:
        '204':
          description: 删除成功
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/view:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Tasks]
      summary: 记录任务已读
      description: |
        成员打开任务详情时，前端调用此接口记录首次查看时间。
        
        - 若已有 viewedAt 记录，则不做任何操作（首次查看语义，不覆盖）。
        - 若记录不存在则创建，写入 viewedAt = now()。
        - 此接口对前端而言是 fire-and-forget，失败不影响用户阅读任务。
      security:
        - userAuth: []
      responses:
        '204':
          description: 成功（无返回体）
  /tasks/{taskId}/state:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Tasks]
      summary: 更新当前用户对任务的个人状态
      description: 更新 tags 和 sortOrder。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskStateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskUserState'

  # ══════════════════════════════════════════════════════
  # SUBMISSIONS
  # ══════════════════════════════════════════════════════
  /tasks/{taskId}/submissions:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Submissions]
      summary: 获取任务的提交列表（管理员视图）
      description: |
        仅 OWNER 和 ADMIN 可访问。
        
        返回班级所有成员的提交情况，**包括未提交的成员**。
        查询从 class_members 出发做 LEFT JOIN submissions，保证名单完整。
      security:
        - userAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubmissionListRow'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/submissions/me:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Submissions]
      summary: 获取当前用户对此任务的提交
      security:
        - userAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
        '404':
          description: 当前用户尚未提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Submissions]
      summary: 创建或更新当前用户的提交
      description: |
        Upsert 语义：若提交记录不存在则创建，已存在则更新。
        
        - 创建时写入 firstSubmittedAt，后续更新不修改此字段。
        - 更新时 lastUpdatedAt 自动刷新。
        - 附件通过单独的附件上传接口处理，此接口只接受 Markdown 正文。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSubmissionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
        '403':
          description: 提交已被锁定（allowLateSubmission = false 且已过期）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/submissions/{submissionId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: submissionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Submissions]
      summary: 获取单条提交详情（含附件）
      description: |
        仅 OWNER 和 ADMIN 可访问，用于评分页查看单个提交的正文与附件。
      security:
        - userAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 提交不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/submissions/{submissionId}/grade:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: submissionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Submissions]
      summary: 对提交评分
      description: |
        仅 OWNER 和 ADMIN 可操作。
        自动将 reviewerId 设为当前用户，reviewedAt 设为当前时间。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeSubmissionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/submissions/export:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Submissions]
      summary: 导出成绩 CSV
      description: |
        仅 OWNER 和 ADMIN 可操作。
        
        包含所有班级成员，未提交的成员提交相关字段为空。
        
        CSV 列：昵称、学校、学号、班级、任务名称、首次提交时间、最后修改时间、成绩
      security:
        - userAuth: []
      responses:
        '200':
          description: CSV 文件
          content:
            text/csv:
              schema:
                type: string
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="submissions_export.csv"'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ══════════════════════════════════════════════════════
  # ATTACHMENTS
  # ══════════════════════════════════════════════════════
  /tasks/{taskId}/attachments:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Attachments]
      summary: 上传任务附件（任务正文用）
      description: 仅 OWNER 和 ADMIN 可操作。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentMeta'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}/submissions/me/attachments:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Attachments]
      summary: 上传提交附件
      description: |
        为当前用户的提交上传附件。
        若提交记录尚未创建，上传时自动创建。
        旧附件在上传新附件后由服务端删除（MinIO + DB）。
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentMeta'
...
  /tasks/{taskId}/submissions/rename:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Attachments]
      summary: 批量重命名提交附件
      description: |
        仅 OWNER 和 ADMIN 可操作。
        
        按模板 `{班级名}_{昵称}_{学号}_{原文件名}` 更新每条提交附件的 renamedFile 字段。
        MinIO 对象键不变，只更新展示名。
      security:
        - userAuth: []
      responses:
        '204':
          description: 成功

  # ══════════════════════════════════════════════════════
  # ADMIN
  # ══════════════════════════════════════════════════════
...